Move calculation of lines to adjust to financial processor
This is a straight move but there are areas to consolidate post move
